Fixing the hardfault
Hardfault was trigged by the strcat trying to access at the same time that the DMA might rewrite the values. Values are then not correct but should'nt stop the system.
Also dma has been set as circular, as if too many characterers are incoming without delay then it shouldn't overflow the buffer.